WIP; Developed a page to show details of a match, breaking down the ELO score calculations by competition. This identified a bug where the opponent calculation was rounding to integer and getting the wrong Expected Score, which is now fixed. Added a create view for match and will continue development towards a user being able to add a match

+++ PingPongLeague/Calculators/ELORatingCalculator.cs
@@ -17,7 +17,7 @@ namespace PingPongLeague.Calculators
 
 			ratings.OpponentRating = opponentRating;
 
-			ratings.OpponentTransformedRating = Math.Pow(10, (ratings.OpponentRating / ELO_N_VALUE));
+			ratings.OpponentTransformedRating = Math.Pow(10, ((double)ratings.OpponentRating / ELO_N_VALUE));
 
 			ratings.ExpectedScore = ratings.TransformedRating / (ratings.TransformedRating + ratings.OpponentTransformedRating);
